lk2nd: display: Add continuous splash display implementation
Most devices leave the panel powered on to keep the splash screen
visible while the kernel is loading. This allows having a very
simple display implementation in the secondary bootloader by using
the memory region that was allocated for the continuous splash.

The concept was originaly borrowed from efidroid:
https://github.com/efidroid/bootloader_lk/tree/common/shared/lib/2ndstage_display
but the code has almost nothing in common at this point.

This commit is a squashed version of significant rework over the time
to make it work for various display configurations and platforms.
See the old lk2nd "master" branch for details.