lk2nd: Implement more sophisticated DTB override mechanism
So far the code was trying to adjust the board information in some
funky way so that QCOM's DTB selection algorithm would possibly
select the same DTB as the stock bootloader did. But this isn't
particularly reliable because all the masks etc do not apply properly
to the board-ids etc used by other vendors in their stock bootloaders.

So instead, override QCOM's DTB selection algorithm in that case,
to only match exactly the same board-id that the stock bootloader
selected for lk2nd.

Also use dev_tree_compatible(...) for parsing the board-id in the
lk2nd dts instead so we also handle older variants like only qcom,msm-id.

Change-Id: Ib5244e2fe04f5b02adde876ac85640eb6e26fd97