display_menu: Make it clear that unknown models should be fixed
At the moment it seems like not having a model displayed is just fine,
because everything often seems working. But in fact it's bad, especially
if we want to have some automatic DTB selection based on this in the future.

So make it clear that unknown models should be fixed by adding proper
entries in the DTBs.

Change-Id: I9bb9613510ce4e9f8c32c64cfe302feebeebfc94