arch: arm: crt0: Generate Linux ARM32 header
Qualcomm's bootloader relies on a header in the assembly of the
ARM32 Linux zImage that tells the bootloader where the kernel image
ends. This is used to figure out where the appended DTB starts.

The only alternative for this is compressing the kernel image with
GZIP but that works only for newer ARM64 bootloaders (since ARM32
Linux can decompress itself without help from the bootloader).