target: msm8916: Abuse firmware bugs to load qhypstub
It's possible to abuse some scm calls that don't check if the write
destination is inside the hyp memory and corrupt the boundary check for
the hypervisor call that allows mapping memory to EL1. Then we can just
write arbitrary code into the hyp memory and it will replace the
hypervisor.

Change-Id: I6fbfbd216a7c0cabdf2b0588291bbe5a6dc01444