makefile: make sure lk2nd version tag is regenerated on build
lk2nd version define was passed as -D CFLAGS argument. This means that
if git thag has changed but the file that prints it wasn't rebuilt, the
version will stay. Move the define to config.h so make properly rebuilds
everything on the tag change.

Add it into the generation script because it changes the format of the
passed values to the configs vriable.