lk2nd: smb1360: wt88047: Allow fallback to 0 Common (2200 mAh)
It seems like some fake replacement batteries might use this one
on the downstream kernel and it would still be nice to have them
working on mainline as well. Especially because it's required for USB
at the moment.

So, uncomment the fallback code again but make it more obvious by
adding a special battery type that displays as WARNING on the lk2nd
screen.