lk2nd: Add hack to reset SMD RPM channel back to OPENING state
Some devices communicate with RPM from LK. In that case, the SMD RPM
channel is temporarily opened and closed again, which will change the
channel state to CLOSING. This confuses the mainline Linux driver and
prevents using the RPM there.

Ideally this should be fixed in Linux. However, different fixes have
been proposed and so far no proper solution has been found. As a simple
workaround it seems also be enough to update the RPM channel state
from lk2nd to reset it back to OPENING.

Note that in general this is VERY BAD since this field is supposed
to be controlled by the RPM and not us. However, it seems to work
fine and is a simple workaround that can be enabled conditionally.

Change-Id: Ifcc854eaec035f1424588f3184e71f8c7ff57e8a