aboot: memmove() kernel/ramdisk after looking for appended DTBs
At the moment, booting ARM32 kernels with CONFIG_ARM_APPENDED_DTB=y
results in a hang if the DTB is indeed appended to the kernel image
given to LK. This happens because both Linux and LK still find it
appended to the kernel image, so Linux will prefer the appended DTB
over the correct DTB passed by LK.

To avoid this, dev_tree_appended() clears the DTB magic of the
appended DTB. However, this has no effect if this is made on
a different copy of the kernel (because the kernel was already copied
to the right location before looking for an appended DTB).

To fix this, memmove() the kernel and ramdisk after looking for
an appended DTB. With this, the DTB magic for the appended DTB will
be correctly cleared in the final kernel image. This allows booting
ARM32 kernels with CONFIG_ARM_APPENDED_DTB=y.