aboot: Make lk2nd reset retry count of booted slot
The stock bootloader decrements retry count by 1 when booting lk2nd, so
set it back to the max 7 as we've booted successfully. When lk2nd boots
an image it will also decrement the value again anyways, afterwards user
space is supposed to set this, but there is also currently no such tool
outside of Android.