target: display: cont_splash_mdp5: Avoid stupid delays for cmd mode
The goal for the code is to avoid triggering the refresh too often.
In that case display updates would overlap and cause screen corruption.
However, at the moment this is implemented in a way that causes a lot
of lags because every single update of the screen needs to wait 20 ms.

Instead, let's make use of a thread that waits for necessary display
refresh and then sleeps for 20 ms. That way the code filling the
display can continue as normally in the meantime.