lk2nd: smp: spin-table: Check CPU "enable-method" first
There might be kernels that can handle the CPU boot using custom
"enable-method"s themselves. This is primarily ARM32 Linux but also
older (modified) ARM64 Linux kernels that do not make use of spin-table
yet. In this case we should not boot the CPUs using spin-table but
rather let the kernel handle it later.

So far this situation is partially covered by checking if the /psci
node is already disabled. However, this is only the case on old MSM8916
device trees. Other device trees for ARM32 SoCs like MSM8974 or SoCs
without any PSCI firmware like MSM8939 might not define any /psci node.

To handle this situation reliably, look at the "enable-method" of one
of the CPUs instead and see if we are supposed to apply spin-table.
Drop the /psci check again because it turned out to be unreliable.