platform: msm_shared: i2c_qup: Unset dev_addr in qup_i2c_deinit()
The i2c_qup driver currently only supports one active I2c controller
at the same time, because a global variable is used to point to the
struct for use in the interrupt handler. While the struct is freed
in qup_i2c_deinit(), the global dev_addr variable is not unset, which
means that it is impossible to register another I2C controller later.

Make sure to unset the variable to fix this. (Ideally the driver would
support multiple active controllers, though.)