dev: fbcon: Improve display menu scaling
Right now the display menu scales quite differently on 720p and 1080p.
It's too big on 1080p, and okay-ish on 720p. The comments in fbcon.h
are quite confusing since they ignore the character width of the font.
Let's make this consistent by:

  - Increasing the amount of characters by line
  - Rounding down/up as needed to have the closest scaling factor

A few examples for the scaling factor calculation:

 Before:
   720p:  720/6/45 = 2.66666... = ~ 2
  1080p: 1080/6/45 = 4.00000... = ~ 4
 After:
   720p:  720/6/60 = 2
  1080p: 1080/6/60 = 3