platform: msm_shared: debug: Initialize log buffer at runtime
Using a static initializer for the log buffer (mainly to set the magic
cookie) requires placing it entirely in the data segment of the binary.
This means that the binary size grows when increasing the size of the
log buffer, even if most of the reserved space would just be zero.

Place the log buffer in the BSS instead and add a debug_init() call
to finish initialization during start up. This makes the final binary
size independent of the log buffer size.