app: aboot: Publish partition-size/type for flash as well
At the moment the "partition-size:" and "partition-type:" fastboot
variables are only published when using an eMMC. However, the flash
also has a partition table (although a quite different one) so similar
information is available there as well.

Add code similar to the publish_getvar_partition_info() for the flash
ptable. This is needed e.g. for "fastboot fetch" to work correctly,
since it needs to know the partition size.