msm8974: fix detection of Pro variants
There are two functions to detect if the SoC is a Pro variant or
the more specific Pro-AC variant.

The implementation of the two functions is seemingly swapped, as
the function containing only the AC models is obviously the one
detecting for the AC variant.

Fix the functions by swapping them around.

Signed-off-by: Alexandre Messier <alex@me.ssier.org>